+++ packages/coachmark/src/Coachmark.ts
@@ -19,6 +19,7 @@ import { property } from '@spectrum-web-components/base/src/decorators.js';
 import { when } from '@spectrum-web-components/base/src/directives.js';
 import coachmarkStyles from './coachmark.css.js';
 import chevronStyles from '@spectrum-web-components/icon/src/spectrum-icon-chevron.css.js';
+import chevronIconOverrides from '@spectrum-web-components/icon/src/icon-chevron-overrides.css.js';
 import '@spectrum-web-components/icons-ui/icons/sp-icon-chevron200.js';
 import { Popover } from '@spectrum-web-components/popover';
 import { join } from '@spectrum-web-components/base/src/directives.js';
@@ -29,7 +30,6 @@ import type { CoachmarkItem } from './CoachmarkItem.js';
 import '@spectrum-web-components/asset/sp-asset.js';
 import '@spectrum-web-components/button/sp-button.js';
 import '@spectrum-web-components/button-group/sp-button-group.js';
-import '@spectrum-web-components/quick-actions/sp-quick-actions.js';
 
 /**
  * @element sp-coachmark
@@ -43,7 +43,12 @@ import '@spectrum-web-components/quick-actions/sp-quick-actions.js';
  */
 export class Coachmark extends Popover {
     public static override get styles(): CSSResultArray {
-        return [...super.styles, coachmarkStyles, chevronStyles];
+        return [
+            ...super.styles,
+            coachmarkStyles,
+            chevronStyles,
+            chevronIconOverrides,
+        ];
     }
     @property({ type: Object })
     public item?: CoachmarkItem;